Rework the parenting of sub-parts.
Much thanks to Kreuzung for the initial work of finding and fixing the
bugs. I just made it work with the original code-base.

This fixes not only the visually broken joints when hinges are mounted
upside down, but also the displaced sub-parts when loading a game saved
with transformed sub-parts.